+++ tests/js/index.test.js
@@ -27,3 +27,4 @@ require('./landscape.test.js');
 require('./theme-animation.test.js');
 require('./hijri.test.js');
 require('./laicite-update.test.js');
+require('./responsive.test.js');